+++ app/src/main/java/com/goalmaster/task/data/source/LocalTaskDataSource.kt
@@ -1,16 +1,22 @@
-package com.goalmaster.task
+package com.goalmaster.task.data.source
 
 import com.goalmaster.Result
 import com.goalmaster.goal.data.source.LocalGoalDataSource
-import com.goalmaster.plan.data.source.PlanDao
+import com.goalmaster.plan.data.entity.PlanTaskStatus
+import com.goalmaster.plan.data.source.PlanTaskDataSource
 import com.goalmaster.task.data.entity.Task
+import com.goalmaster.task.data.entity.TaskState
 import com.goalmaster.task.data.entity.TaskWithData
+import com.goalmaster.task.data.source.TaskDao
+import com.goalmaster.task.data.source.TaskDataSource
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.withContext
 
 class LocalTaskDataSource(
     private val taskDao: TaskDao,
+    private val goalDataSource: LocalGoalDataSource,
+    private val planTaskDataSource: PlanTaskDataSource,
     private val ioDispatcher: CoroutineDispatcher
 ) : TaskDataSource {
 
@@ -30,12 +36,8 @@ class LocalTaskDataSource(
 
     override suspend fun getTask(taskId: Long): Result<Task> = withContext(ioDispatcher) {
         return@withContext try {
-            val task = taskDao.findById(taskId)
-            if (task == null) {
-                Result.Error(Exception("Task not found!"))
-            } else {
-                Result.Success(task)
-            }
+            val task = taskDao.findById(taskId)  ?: throw Exception("Task not found")
+            Result.Success(task)
         } catch (exception: Exception) {
             Result.Error(exception)
         }
@@ -60,8 +62,14 @@ class LocalTaskDataSource(
 
     override suspend fun deleteTask(taskId: Long): Result<Unit> = withContext(ioDispatcher) {
         return@withContext try {
-            val task = taskDao.deleteById(taskId)
-            Result.Success(task)
+            val task = taskDao.findById(taskId) ?: throw Exception("Task not found")
+            if (task.state == TaskState.PLANNED) {
+                unPlanTask(task)
+                task.state = TaskState.UNPLANNED
+            }
+            if (task.state != TaskState.UNPLANNED) throw Exception("Illegal state")
+            taskDao.deleteById(taskId)
+            Result.Success(Unit)
         } catch (exception: Exception) {
             Result.Error(exception)
         }
@@ -71,16 +79,30 @@ class LocalTaskDataSource(
         return taskDao.observeTasksForPlan()
     }
 
-    override suspend fun updateTaskState(task: Task, state: TaskState): Result<Unit> = withContext(ioDispatcher) {
+    override suspend fun updateTaskState(id: Long, state: TaskState): Result<Unit> = withContext(ioDispatcher) {
         return@withContext try {
-            task.state = state
-            if (task.state == TaskState.DONE) {
-                taskDao.markPlanTaskCompletedByTaskId(task.id)
+            val task = taskDao.findById(id) ?: throw Exception("Task not found")
+            when (state) {
+                TaskState.DONE -> doneTask(task)
+                TaskState.PLANNED -> {}
+                TaskState.UNPLANNED -> unPlanTask(task)
             }
+            task.state = state
             taskDao.insert(task)
             Result.Success(Unit)
         } catch (exception: Exception) {
             Result.Error(exception)
         }
+    }
+
+    override suspend fun unPlanTask(task: Task): Result<Unit> {
+        planTaskDataSource.updateStatusByTaskId(task.id, PlanTaskStatus.FAILED)
+        return Result.Success(Unit)
+    }
+
+    override suspend fun doneTask(task: Task): Result<Unit> {
+        planTaskDataSource.updateStatusByTaskId(task.id,PlanTaskStatus.COMPLETED)
+        goalDataSource.updateCompletedUnits(task.goalId, task.unitSize)
+        return Result.Success(Unit)
     }
 }
